+++ src/main/java/com/example/easyexpressbackend/config/RabbitMQConfig.java
@@ -32,17 +32,4 @@ public class RabbitMQConfig {
 
         return simpleMessageListenerDeliveredEmailContainer;
     }
-
-//    @Bean
-//    MessageListenerContainer messageListenerDeliveredEmailContainer2(
-//            ConnectionFactory connectionFactory,
-//            DeliveredEmailWorker2 deliveredEmailWorker2
-//    ) {
-//        SimpleMessageListenerContainer simpleMessageListenerDeliveredEmailContainer = new SimpleMessageListenerContainer();
-//        simpleMessageListenerDeliveredEmailContainer.setConnectionFactory(connectionFactory);
-//        simpleMessageListenerDeliveredEmailContainer.setQueues(deliveredEmailQueue());
-//        simpleMessageListenerDeliveredEmailContainer.setMessageListener(deliveredEmailWorker2);
-//
-//        return simpleMessageListenerDeliveredEmailContainer;
-//    }
 }
