+++ src/main/java/com/example/easyexpressbackend/controller/tracking/PublicTrackingController.java
@@ -2,6 +2,7 @@ package com.example.easyexpressbackend.controller.tracking;
 
 import com.example.easyexpressbackend.response.tracking.TrackingAShipmentResponse;
 import com.example.easyexpressbackend.service.ShipmentService;
+import com.example.easyexpressbackend.service.TrackingService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -11,15 +12,15 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping(path = "/api/public/trackings")
 public class PublicTrackingController {
-    private final ShipmentService shipmentService;
+    private final TrackingService trackingService;
 
     @Autowired
-    public PublicTrackingController(ShipmentService shipmentService) {
-        this.shipmentService = shipmentService;
+    public PublicTrackingController(TrackingService trackingService) {
+        this.trackingService = trackingService;
     }
 
     @GetMapping({"/",""})
     public TrackingAShipmentResponse listTrackingOfShipment(@RequestParam("shipment") String shipmentNumber){
-        return shipmentService.trackingAShipment(shipmentNumber);
+        return trackingService.trackingAShipment(shipmentNumber);
     }
 }
