+++ src/main/java/com/example/easyexpressbackend/controller/tracking/TrackingController.java
@@ -2,7 +2,7 @@ package com.example.easyexpressbackend.controller.tracking;
 
 import com.example.easyexpressbackend.dto.tracking.AddTrackingDto;
 import com.example.easyexpressbackend.response.tracking.TrackingPrivateResponse;
-import com.example.easyexpressbackend.service.ShipmentService;
+import com.example.easyexpressbackend.service.TrackingService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -13,17 +13,18 @@ import org.springframework.web.bind.annotation.RestController;
 @RestController
 @RequestMapping(path = "/api/trackings")
 public class TrackingController {
-    private final ShipmentService shipmentService;
+
+    private final TrackingService trackingService;
 
     @Autowired
-    public TrackingController(ShipmentService shipmentService) {
-        this.shipmentService = shipmentService;
+    public TrackingController(TrackingService trackingService) {
+        this.trackingService = trackingService;
     }
 
 
     @PostMapping({"/",""})
     public TrackingPrivateResponse addTracking(@RequestBody @Valid AddTrackingDto addTrackingDto){
 
-        return shipmentService.addTracking(addTrackingDto);
+        return trackingService.addTrackingContinued(addTrackingDto);
     }
 }
