+++ src/main/java/com/example/easyexpressbackend/utils/Guards.java
@@ -1,0 +1,8 @@
+package com.example.easyexpressbackend.utils;
+
+public class Guards {
+    public static void mustPositive(String argumentName, int value) {
+        if (value <= 0)
+            throw new IllegalArgumentException(argumentName + " should be a positive number");
+    }
+}
