--- src/main/resources/templates/test.html
@@ -1,36 +1,0 @@
-<html>
-<head>
-    <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />
-    <style>
-        /* Add custom classes and styles that you want inlined here */
-    </style>
-</head>
-<body class="bg-light">
-<div class="container">
-    <div class="card my-10">
-        <div class="card-body">
-            <h1 class="h3 mb-2">Bootstrap Email</h1>
-            <h5 class="text-teal-700">Here is a very simple email using Bootstrap Email</h5>
-            <hr>
-            <div class="space-y-3">
-                <p class="text-gray-700">Click the "Render" button at the top of the page every time you want to rebuild the email.</p>
-                <p class="text-gray-700">
-                    We hope you enjoy using Bootstrap Email. This is our Online Editor that allows you to edit emails
-                    and have them render directly in the browser. The outputted HTML will be in the "HTML" tab in the top
-                    and that is the HTML you want to use to send emails. All the styles and responsive CSS are self contained in the
-                    HTML that is generated making is very portable needing no external CSS files.
-                </p>
-                <p class="text-gray-700">
-                    You can use the "Test" button to email yourself the outputted code to test in your inbox and email client.
-                </p>
-                <p class="text-gray-700">
-                    Check out the <a href="https://bootstrapemail.com/docs/introduction" target="_blank">Documentation</a> for syntax and usage of writing emails with Bootstrap Email.
-                </p>
-            </div>
-            <hr>
-            <a class="btn btn-primary" href="https://app.bootstrapemail.com/templates" target="_blank">Get More Email Templates</a>
-        </div>
-    </div>
-</div>
-</body>
-</html>
